pyproject.toml
[project]
name = "ai-email-triage"
version = "0.1.0"
description = "LangGraph 기반 멀티 에이전트 이메일 트리아지 & 자동 응답 시스템"
requires-python = ">=3.11"

dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langsmith>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "streamlit>=1.40.0",
    "typer>=0.12.0",
    "rich>=13.9.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
triage = "cli.main:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
